Move sanity client file to the root of the project
